feat(control): ajout du lecteur MIDI et du bouton de téléchargement après génération
